refactor: upgrade request to TS
